+++ tests/js/object-id-tests.js
@@ -49,12 +49,12 @@ module.exports = {
         }
         
         return Realm.Sync.User.register('http://localhost:9080', uuid(), 'password').then(user => {
-            const config = { sync: { user, url: 'realm://localhost:9080/~/myrealm' },
+            const config = user.createConfiguration({ sync: { url: 'realm://localhost:9080/~/myrealm' },
                              schema: [{ name: 'IntegerPrimaryKey', properties: { int: 'int?' }, primaryKey: 'int' },
                                       { name: 'StringPrimaryKey', properties: { string: 'string?' }, primaryKey: 'string' },
                                       { name: 'NoPrimaryKey', properties: { string: 'string' }},
                                      ],
-                           };
+                           });
             return Realm.open(config).then(realm => {
                 var integer, nullInteger;
                 var string, nullString;