+++ tests/js/session-tests.js
@@ -116,7 +116,7 @@ module.exports = {
                 // Let the error handler trigger our checks when the access token was refreshed.
                 postTokenRefreshChecks._notifyOnAccessTokenRefreshed = accessTokenRefreshed;
 
-                const config = { sync: { user, url: 'realm://localhost:9080/~/myrealm', error: postTokenRefreshChecks } };
+                const config = user.createConfiguration({ sync: { url: 'realm://localhost:9080/~/myrealm', error: postTokenRefreshChecks, fullSynchronization: true } });
                 const realm = new Realm(config);
                 const session = realm.syncSession;
                 TestCase.assertInstanceOf(session, Realm.Sync.Session);
@@ -148,7 +148,7 @@ module.exports = {
                 let successCounter = 0;
 
                 config = {
-                    sync: { user, url: `realm://localhost:9080/~/${realmName}` },
+                    sync: { user, url: `realm://localhost:9080/~/${realmName}`, fullSynchronization: true },
                     schema: [{ name: 'Dog', properties: { name: 'string' } }],
                 };
 
@@ -184,7 +184,7 @@ module.exports = {
                 let successCounter = 0;
 
                 config = {
-                    sync: { user, url: `realm://localhost:9080/~/${realmName}` },
+                    sync: { user, url: `realm://localhost:9080/~/${realmName}`, fullSynchronization: true },
                     schema: [{ name: 'Dog', properties: { name: 'string' } }],
                     schemaVersion: 1,
                 };
@@ -226,7 +226,7 @@ module.exports = {
                 let successCounter = 0;
 
                 let config = {
-                    sync: { user, url: `realm://localhost:9080/~/${realmName}` },
+                    sync: { user, url: `realm://localhost:9080/~/${realmName}`, fullSynchronization: true },
                     schema: [{ name: 'Dog', properties: { name: 'string' } }],
                 };
                 return new Promise((resolve, reject) => {
@@ -277,7 +277,7 @@ module.exports = {
                 let successCounter = 0;
 
                 let config = {
-                    sync: { user, url: `realm://localhost:9080/~/${realmName}` }
+                    sync: { user, url: `realm://localhost:9080/~/${realmName}`, fullSynchronization: true }
                 };
                 return new Promise((resolve, reject) => {
                     Realm.openAsync(config, (error, realm) => {
@@ -372,7 +372,7 @@ module.exports = {
     testErrorHandling() {
         return Realm.Sync.User.register('http://localhost:9080', uuid(), 'password').then(user => {
             return new Promise((resolve, _reject) => {
-                const config = { sync: { user, url: 'realm://localhost:9080/~/myrealm' } };
+                const config = user.createConfiguration({ sync: { url: 'realm://localhost:9080/~/myrealm' } });
                 config.sync.error = (sender, error) => {
                     try {
                         TestCase.assertEqual(error.message, 'simulated error');
@@ -405,7 +405,7 @@ module.exports = {
             .then(user => {
                 let config = {
                     schema: [schemas.ParentObject, schemas.NameObject],
-                    sync: { user, url: `realm://localhost:9080/~/${realmName}` }
+                    sync: { user, url: `realm://localhost:9080/~/${realmName}`, fullSynchronization: true }
                 };
                 return Realm.open(config)
             }).then(realm => {
@@ -744,8 +744,8 @@ module.exports = {
                     sync: {
                         user: user,
                         url: `realm://localhost:9080/default/__partial/`,
-                        partial: true,
-                        _disablePartialSyncUrlChecks: true
+                        _disableQueryBasedSyncUrlChecks: true,
+                        fullSynchronization: false,
                     }
                 };
                 const realm = new Realm(config1);
@@ -758,7 +758,7 @@ module.exports = {
                 sync: {
                     user: user,
                     url: `realm://localhost:9080/default/__partial/`,  // <--- not allowed URL
-                    partial: true,
+                    fullSynchronization: false,
                 }
             };
             TestCase.assertThrows(() => new Realm(config2));
@@ -769,7 +769,7 @@ module.exports = {
                 sync: {
                     user: user,
                     url: 'realm://localhost:9080/~/default',
-                    partial: false, // <---- calling subscribe should fail
+                    fullSynchronization: true, // <---- calling subscribe should fail
                     error: (session, error) => console.log(error)
                 },
                 schema: [{ name: 'Dog', properties: { name: 'string' } }]
@@ -798,7 +798,7 @@ module.exports = {
                     defaultRealmInvalidArguments();
 
                     return new Promise((resolve, reject) => {
-                        let config = Realm.automaticSyncConfiguration();
+                        let config = Realm.Sync.User.current.createConfiguration();
                         config.schema = [{ name: 'Dog', properties: { name: 'string' } }];
                         Realm.deleteFile(config);
 
@@ -874,7 +874,7 @@ module.exports = {
         return Realm.Sync.User.register('http://localhost:9080', uuid(), 'password').then(user => {
             return new Promise((resolve, _reject) => {
                 var realm;
-                const config = { sync: { user, url: 'realm://localhost:9080/~/myrealm' } };
+                const config = user.createConfiguration({ sync: { url: 'realm://localhost:9080/~/myrealm' } });
                 config.sync.error = (sender, error) => {
                     try {
                         TestCase.assertEqual(error.name, 'ClientReset');