+++ tests/js/user-tests.js
@@ -130,7 +130,8 @@ module.exports = {
       }).then((user => {
           assertIsUser(user);
           // Can we open a realm with the logged-in user?
-          const realm = new Realm({ sync: { user: user, url: 'realm://localhost:9080/~/test' } });
+          const config = user.createConfiguration({ sync: { url: 'realm://localhost:9080/~/test' }});
+          const realm = new Realm(config);
           TestCase.assertInstanceOf(realm, Realm);
           realm.close();
       }))
@@ -143,7 +144,7 @@ module.exports = {
       return Realm.Sync.User.authenticate('http://localhost:9080', 'password', { username: username, password: 'password' });
     }).then((user => {
       assertIsUser(user);
-      const realm = new Realm({ sync: { user: user, url: 'realm://localhost:9080/~/test' } });
+      const realm = new Realm(user.createConfiguration({ sync: { url: 'realm://localhost:9080/~/test' } }));
       TestCase.assertInstanceOf(realm, Realm);
       realm.close();
     }))
@@ -317,6 +318,58 @@ module.exports = {
       TestCase.assertEqual(e.type, "https://realm.io/docs/object-server/problems/unknown-account");
     }).then(account => { if (account) { throw new Error("Retrieving nonexistent account should fail"); }});
   },
+
+  testCreateConfiguration_defaultConfig() {
+      const username = uuid();
+      return Realm.Sync.User.register('http://localhost:9080', username, 'password').then((user) => {
+          let config = user.createConfiguration();
+          TestCase.assertEqual(config.sync.url, "realm://localhost:9080/default");
+          TestCase.assertUndefined(config.sync.partial);
+          TestCase.assertFalse(config.sync.fullSynchronization);
+      });
+  },
+
+  testCreateConfiguration_useOldConfiguration() {
+      const username = uuid();
+      return Realm.Sync.User.register('http://localhost:9080', username, 'password').then((user) => {
+          let config = user.createConfiguration({ sync: { url: 'http://localhost:9080/other_realm', partial: true }});
+          TestCase.assertEqual(config.sync.url, 'http://localhost:9080/other_realm');
+          TestCase.assertUndefined(config.sync.fullSynchronization);
+          TestCase.assertTrue(config.sync.partial);
+      });
+  },
+
+  testCreateConfiguration_settingPartialAndFullSynchronizationThrows() {
+      const username = uuid();
+      return Realm.Sync.User.register('http://localhost:9080', username, 'password').then((user) => {
+          TestCase.assertThrowsContaining(() => {
+                  let config = { 
+                    sync: { 
+                      url: 'http://localhost:9080/~/default',
+                      partial: true, 
+                      fullSynchronization: false 
+                    } 
+                  };
+                  user.createConfiguration(config);
+          }, "'partial' and 'fullSynchronization' were both set. 'partial' has been deprecated, use only 'fullSynchronization'");
+      });
+  },
+    
+  testOpen_partialAndFullSynchronizationSetThrows() {
+      const username = uuid();
+      return Realm.Sync.User.register('http://localhost:9080', username, 'password').then((user) => {
+          TestCase.assertThrowsContaining(() => {
+              new Realm({
+                  sync: {
+                    user: user,
+                    url: 'http://localhost:9080/~/default',
+                    partial: false,
+                    fullSynchronization: true
+                  }
+              })
+          }, "'partial' and 'fullSynchronization' were both set. 'partial' has been deprecated, use only 'fullSynchronization'");
+      });
+  }
 
   /* This test fails because of realm-object-store #243 . We should use 2 users.
   testSynchronizeChangesWithTwoClientsAndOneUser() {